Group_12/requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cssselect==1.0.3
cycler==0.10.0
Django==2.1.7
django-autocomplete-light==3.3.2
django-ckeditor==5.6.1
django-cors-headers==2.5.2
django-filter==2.1.0
django-js-asset==1.2.2
djangorestframework==3.9.2
idna==2.8
isort==4.3.15
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.3
Markdown==3.0.1
markdown2==2.3.7
matplotlib==3.0.3
mccabe==0.6.1
numpy==1.16.2
Pillow==5.4.1
psycopg2==2.7.7
pylint==2.3.1
pyparsing==2.3.1
pyquery==1.4.0
python-dateutil==2.8.0
pytz==2018.9
readtime==1.1.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
